+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #284 Update to bootique-modules-parent 2.0
 * #286 Update to bootique-modules-parent 2.0.1 
 * #287 bootique-test-junit5: manage static BQRuntime fields 
+* #288 Ambiguos "Bootique.module(..)" methods
 
 ## 1.1